+++ messages/src/lib.rs
@@ -1,9 +1,7 @@
 pub(crate) mod api;
-pub(crate) mod chunking;
 pub(crate) mod message;
 pub(crate) mod protocol;
 
 pub use api::ApplicationAPI;
-pub use chunking::{MessageChunker, SimpleChunker};
 pub use message::Message;
 pub use protocol::{DataProtocol, TransmissionBlock};
