+++ myceli/Cargo.toml
@@ -17,6 +17,7 @@ tokio = "1"
 toml = "0.7"
 tracing = "0.1"
 tracing-subscriber = "0.3"
+transports = { path = "../transports" }
 
 [dev-dependencies]
 assert_fs = "1"
