+++ myceli/src/lib.rs
@@ -2,5 +2,3 @@ pub mod config;
 mod handlers;
 pub mod listener;
 pub mod shipper;
-pub mod transport;
-pub mod udp_transport;
