+++ transports/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "transports"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+cid = { version = "0.10" }
+messages = { path = "../messages" }
+parity-scale-codec = { version = "3.0.0", default-features = false, features = ["derive", "std"] }
+parity-scale-codec-derive = "3.1.3"
+rand = "0.8.5"
+serde = { version = "1", features = ["derive"] }
+tracing = "0.1"
