+++ transports/src/lib.rs
@@ -1,10 +1,13 @@
+mod chunking;
+mod udp_chunking;
+mod udp_transport;
+
 use anyhow::Result;
 use messages::Message;
 
 pub trait Transport: Send + Sync {
-    fn new(listen_addr: &str, mtu: u16) -> Result<Self>
-    where
-        Self: Sized;
     fn receive(&self) -> Result<(Message, String)>;
     fn send(&self, msg: Message, addr: &str) -> Result<()>;
 }
+
+pub use udp_transport::UdpTransport;
